Create GUI
Hello, 
So I am new to GitHub and looking to just experiment with different Java concepts for practice. I was looking to get some feed back about this code. First, I am having trouble producing a GUI on my computer that draws the triangle. I can get the GUI screen, but it will not draw anything inside the GUI. I am sure that it is a simple issue but I am still fairly new and need some help.  So I have read a little about the Floyd Triangle and kind of understand the concept, but when I run the original file I notice that when you put in a number over two digits that there appears to be a point at which the numbers form a plane and then continue to the next triangle. Is there an explanation for this? 
PS(I am currently using netbeans as my development environment. I don't know if that would affect anything but figured I would mention it.)